+++ src/app/core/+store/products/products.selectors.ts
@@ -1,16 +1,16 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
 
-import { ProductsState } from './products.state';
+import { ProductsState, productAdapter } from './products.state';
 import { Product } from './../../../products/models/product.model';
 import { getRouterState } from './../router';
 
 
 export const getProductsState = createFeatureSelector<ProductsState>('products');
 
-export const getProductsData = createSelector(
-  getProductsState,
-  (state: ProductsState) => state.data
-);
+export const {
+  selectEntities: getProductsEntities,
+  selectAll: getProductsData
+} = productAdapter.getSelectors(getProductsState);
 
 export const getProductsError = createSelector(
   getProductsState,
@@ -18,15 +18,13 @@ export const getProductsError = createSelector(
 );
 
 export const getProductByUrl = createSelector(
-  getProductsData,
+  getProductsEntities,
   getRouterState,
   (products, router): Product => {
     const productID = router.state.params.productID;
 
     if (productID) {
-      return products.find(product =>
-        product.id === +productID
-      );
+      return products[productID];
     } else {
       return new Product();
     }