+++ src/app/core/+store/products/products.state.ts
@@ -1,11 +1,19 @@
+import {
+  createEntityAdapter,
+  EntityState,
+  EntityAdapter } from '@ngrx/entity';
+
 import { Product } from './../../../products/models/product.model';
 
-export interface ProductsState {
-  data: ReadonlyArray<Product>;
+
+export interface ProductsState extends EntityState<Product> {
   readonly error: Error | string;
 }
 
-export const initialProductsState: ProductsState = {
-  data: [],
-  error: null
-};
+export const productAdapter: EntityAdapter<Product> =
+  createEntityAdapter<Product>();
+
+export const initialProductsState: ProductsState =
+  productAdapter.getInitialState({
+    error: null
+  });